use ReservationPriceCalculator
